requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.9.0
blis>=0.7.0
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
fastai>=1.0
fastapi==0.115.12
fastcore==1.7.29
fastdownload==0.0.7
fastprogress==1.0.3
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.2
groovy==0.1.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.12
networkx==3.4.2
numpy>=1.21.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow>=8.4.0
preshed==3.0.9
pydantic==2.11.1
pydantic_core==2.33.0
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
ruff==0.11.2
safehttpx==0.1.6
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy>=3.0.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.46.1
sympy==1.13.1
thinc==8.3.4
threadpoolctl==3.6.0
tomlkit==0.13.2
torch>=2.0.0
torchvision>=0.15.0
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
weasel==0.4.1
websockets==15.0.1
wrapt==1.17.2
flask>=3.0.2
flask-cors>=4.0.0
pillow>=10.2.0
fastai>=2.7.14
torch>=2.2.1
